default format set to short in xml2struct
svn path=/trunk/matlab/; revision=8234


Former-commit-id: 6ec9d76212fc28c6840f3289339ddaabd32d2030

+++ applications/DelftDashBoard/general/fileio/xml2struct.m
@@ -85,7 +85,7 @@ function s=xml2struct(varargin)
 
 filename=varargin{1};
 includeattributes=1;
-structuretype='long';
+structuretype='short';
 includeroot=0;
 
 for ii=1:length(varargin)
